refactor(sync): separate authority and replica domains
