+++ config/architecture-budgets.json
@@ -8,7 +8,6 @@
     "crates/connect-agent/src/server.rs": 1967,
     "crates/connect-mirror/src/lib.rs": 1915,
     "packages/sync/src/adoption.ts": 1127,
-    "packages/sync/src/index.ts": 1116,
     "packages/sync/src/mirror.ts": 1255,
     "services/server/src/legacy-baseline.ts": 1114
   }
