+++ packages/sync/src/sync-error.ts
@@ -1,0 +1,5 @@
+export class SyncError extends Error {
+  constructor(public readonly code: string, message: string) {
+    super(message);
+  }
+}
